rename marginal module into marginal_distributions
move QuantitativeAttributes to bhepop2.sources.base
